refactor: move visualization workflows into package (#269)
