+++ visualization/application/__init__.py
@@ -1,1 +1,29 @@
 """Application services and ports for visualization workflows."""
+
+from .background_map_controller import (
+    BackgroundMapController,
+    LoadBackgroundRequest,
+    LoadBackgroundResult,
+)
+from .layer_gateway import LayerGateway
+from .visual_apply import (
+    ApplyVisualizationRequest,
+    BackgroundConfig,
+    LayerRefs,
+    VisualApplyRequest,
+    VisualApplyResult,
+    VisualApplyService,
+)
+
+__all__ = [
+    "ApplyVisualizationRequest",
+    "BackgroundConfig",
+    "BackgroundMapController",
+    "LayerGateway",
+    "LayerRefs",
+    "LoadBackgroundRequest",
+    "LoadBackgroundResult",
+    "VisualApplyRequest",
+    "VisualApplyResult",
+    "VisualApplyService",
+]
